+++ tests/testthat/test-report_spec.R
@@ -75,6 +75,17 @@ test_that("Titles, footnotes, header, and footer limits work as expected.", {
 
 })
 
+test_that("Footnotes traps invalid parameter as expected.", {
+  
+  rpt <- create_report("fork.out")
+  
+  expect_error(footnotes(rpt, align = "error"))
+  expect_error(footnotes(rpt, valign = "error"))
+  expect_error(footnotes(rpt, blank_row = "error"))
+  expect_error(footnotes(rpt, borders = "error"))
+  
+})
+
 test_that("add_content works as expected.", {
   
   rpt <- create_report("fork.out")